refactor : api folder 재배치 및 폴더화
