+++ api/auth/code/getGoogleAuthCode.ts
@@ -1,4 +1,7 @@
-export const getGoogleAuthCode = async (code: string, router: any) => {
+import { logIn } from '../../../store/slice/loginSlice';
+import { AppDispatch } from '../../../store/store';
+
+export const getGoogleAuthCode = async (code: string, router: any, dispatch: AppDispatch) => {
   
   try {
     const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/auth/google?code=${code}`, {
@@ -14,7 +17,8 @@ export const getGoogleAuthCode = async (code: string, router: any) => {
     const data = await response.json();
 
     if (data.code === 2000) {
-      alert(data.message)
+      dispatch(logIn());
+      localStorage.setItem('loggedInPlatform', 'google')
       router.push('/');
     } else if (data.code === 2001) {
       router.push('/sign-up');
