+++ api/auth/helper/getIsNIckNameExist.ts
