+++ api/auth/secure/postSignUpMember.ts
